requirements.txt
# Requirements for test suite
paho-mqtt>=2.0.0
colorama>=0.4.6
